+++ include/stream/vtk_viz.h
@@ -39,7 +39,7 @@ private:
   //! \param surrogate_ptr Pointer to the surrogate to write
   //! \param t_rad         Radial resolution of the generated VTK mesh
   SurrogateVtkWriter(const SurrogateHeatDriver& surrogate_ref,
-                     int t_res,
+                     size_t t_res,
                      const std::string& regions_to_write,
                      const std::string& data_to_write);
 
@@ -71,7 +71,7 @@ private:
   //! \return 1-D array of points translated to a center x,y
   xtensor<double, 1> points_for_pin(double x, double y);
   //! Return connectivity with an offset
-  xtensor<int, 1> conn_for_pin(int offset);
+  xtensor<int, 1> conn_for_pin(size_t offset);
   //! Generate fuel connectivity (axial, radial, res, conn)
   //! \return fuel element connectivity (axial, radial, res, conn)
   xtensor<int, 4> fuel_conn();
@@ -93,7 +93,7 @@ private:
 
   // internal variables/parameters
   const SurrogateHeatDriver& surrogate_; //!< reference to surrogate
-  int radial_res_;                       //!< radial resolution;
+  size_t radial_res_;                    //!< radial resolution;
   VizDataType data_out_;
   VizRegionType regions_out_;
 
@@ -104,12 +104,12 @@ private:
 
   // SECTIONS
   // axial
-  int n_axial_sections_; //!< number of axial sections
-  int n_axial_points_;   //!< number of axial points (planes)
+  size_t n_axial_sections_; //!< number of axial sections
+  size_t n_axial_points_;   //!< number of axial points (planes)
   // radial
-  int n_radial_fuel_sections_; //!< number of radial fuel sections
-  int n_radial_clad_sections_; //!< number of radial cladding sections
-  int n_radial_sections_;      //!< total number of radial SECTIONS
+  size_t n_radial_fuel_sections_; //!< number of radial fuel sections
+  size_t n_radial_clad_sections_; //!< number of radial cladding sections
+  size_t n_radial_sections_;      //!< total number of radial SECTIONS
 
   // POINTS
   size_t n_fuel_points_; //!< number of points needed to represent the fuel regions
@@ -117,10 +117,10 @@ private:
   size_t n_points_;      //!< total number of points in the vtk representation
 
   // PER PLANE VALUES
-  int n_sections_per_plane_;  //!< total number of radial sections in a plane
-  int fuel_points_per_plane_; //!< number of fuel points in a plane
-  int clad_points_per_plane_; //!< number of cladding points in a plane
-  int points_per_plane_;      //!< total number of points in a plane
+  size_t n_sections_per_plane_;  //!< total number of radial sections in a plane
+  size_t fuel_points_per_plane_; //!< number of fuel points in a plane
+  size_t clad_points_per_plane_; //!< number of cladding points in a plane
+  size_t points_per_plane_;      //!< total number of points in a plane
 
   // MESH ELEMENTS
   size_t n_fuel_elements_; //!< number of fuel elements in mesh
